refactor: rename context file and update imports for Explorers and Converters context
