+++ DashAI/front/src/components/notebooks/context/ExplorersAndConvertersContext.jsx
@@ -1,0 +1,21 @@
+import React, { createContext, useContext, useState } from "react";
+
+const ExplorersAndConvertersContext = createContext();
+
+export const useExplorersAndConverters = () =>
+  useContext(ExplorersAndConvertersContext);
+
+export const ExplorersAndConvertersProvider = ({ children }) => {
+  const [explorersAndConverters, setExplorersAndConverters] = useState([]);
+
+  const value = {
+    explorersAndConverters,
+    setExplorersAndConverters,
+  };
+
+  return (
+    <ExplorersAndConvertersContext.Provider value={value}>
+      {children}
+    </ExplorersAndConvertersContext.Provider>
+  );
+};
